src/localcrimesmap/requirements.txt
psycopg2
django-bootstrap4
leaflet
crispy_forms